Make algorithms work with batched updates
Many of the algorithms get stuck on arms, either because they expect
the learner’s counts to be updated from zero, or because it always
selects the first arm with the maximum value (which is just the first
arm, when we initialize an MLELearner object to initially take on some
fixed mean for all arms).  This fixes these issues, and makes sure
every arm is played at least once (in cases where this is the expected
behavior), and randomly selects among best arms if there are multiple
best arms.  With this change, all algorithms should in expectation
sample from all arms with equal probability during the first batch.